execution/vm, execution/state, db/state/execctx: move only the tests that need the heavy imports
vm needed no move at all: TestCreateGas was the only test reaching for
rpchelper/testutil/freezeblocks, and the lighter setup the rest of the file
already uses works there too. In state and execctx the unit tests come back
next to their code.
